generalize to allow usage of SDF models to describe the kinematics
In WDLSSolver, it's a change to a robot_model/robot_model_format
property pair that match kdl_parser support for SDF.

However, ToPosConverter also needs the joint limits. A function
in Utilities handles the conversion from the robot model to
base::JointLimits, and ToPosConverter uses that to be format-
independent.